+++ spec/integration/api/v1/status_spec.rb
@@ -32,8 +32,6 @@ describe Vmpooler::API::V1 do
     let(:current_time) { Time.now }
 
     before(:each) do
-      redis.flushdb
-
       app.settings.set :config, config
       app.settings.set :redis, redis
       app.settings.set :config, auth: false