+++ CHANGELOG.md
@@ -10,7 +10,9 @@ The format is based on
 Tracking in this Changelog began for this project with the tagging of version 0.1.0.
 If you're looking for changes from before this, refer to the project's
 git logs & PR history.
-# [Unreleased](https://github.com/puppetlabs/vmpooler/compare/0.3.0...master)
+# [Unreleased](https://github.com/puppetlabs/vmpooler/compare/0.4.0...master)
+
+# [0.4.0](https://github.com/puppetlabs/vmpooler/compare/0.3.0...0.4.0)
 
 ### Fixed
  - Improve support for configuration via environment variables (POOLER-137)